requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
matplotlib
opencv-python
albumentations
Pillow
fastapi
uvicorn
pydantic
requests
python-multipart
tqdm
pyaml
pandas
onnx
onnxruntime
